Code review and sonar qube changes
